+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to the "winnetoujs-vscode-framework" extension will be docum
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.9]
+
+### Updated
+
+- Reload button to reload all vscode
+
+### Fixed
+
+- parse constructos goto reference when path is set in package.json
+
+### Added
+
+- Error message when try to load strings in project that hasn't translations
+
 ## [1.0.8]
 
 ### Fixed